tests/wpt/tests/signed-exchange/service-workers/sxg-sw-register-after-fallback.tentative.https.html

<!DOCTYPE html>
<title>Service worker registration after fallback</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="../resources/sxg-util.js"></script>
<body>
<script>
promise_test(async (t) => {
  const sxgUrl =
      get_host_info().HTTPS_ORIGIN +
      '/signed-exchange/resources/sxg/register-sw-after-fallback.sxg';
  const message = await openSXGInIframeAndWaitForMessage(t, sxgUrl);
  assert_equals(message.location,
                get_host_info().HTTPS_ORIGIN +
                '/signed-exchange/resources/register-sw-after-fallback.html');
  assert_true(message.is_fallback);

  assert_equals(message.err, undefined);
  assert_equals(message.iframe_body, 'Generated by service worker');
}, 'Service worker registration after fallback');
</script>
</body>
